requirements.txt
langchain
chromadb
streamlit
huggingface_hub
sentence-transformers
python-dotenv
